Show media fallback for web player only when needed
Finding the same origin poilcy via an iframe is not exactly straight
forward. This hack relies on triggering a DOMException error by
accessing the inner content of the iframe. If the iframe is protected by
CORS then a `SecurityError` will be thrown instead of a `DOMException`.

Swollow DOMException and propogate media fallback on SecurityError